# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vaya que valla",
        "version": "0.1"
    },
    "paths": {
...
        "/api/quotes": {
            "get": {
                "tags": [
                    "Quotes"
                ],
                "summary": "List quotes with pagination",
                "operationId": "661fc6b5be9fae62117dfb58a5f1732a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "itemsPerPage",
                        "in": "query",
                        "description": "Items per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort by field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Register quote",
                "operationId": "cbc64f2d5bf1efcfdd3391f2cae01ef5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "status",
                                    "request_date",
                                    "total_amount"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "number",
                                        "maxLength": 20
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "request_date": {
                                        "type": "string"
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "maxLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quote registered successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/quotes/{id}": {
            "get": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Get quote by ID",
                "operationId": "114eb8eff718346161619727bb6ba575",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the quote",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quote found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Quote found"
                                        },
                                        "organization": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Quote not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Quotes"
                ],
                "summary": "Update quote",
                "operationId": "6c013f3af9642f741ef8b76798af233a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the quote to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated quote data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "status",
                                    "request_date",
                                    "total_amount"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "number",
                                        "maxLength": 20
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "request_date": {
                                        "type": "string"
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "maxLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quote updated successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT con el formato 'Bearer {token}'.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}